# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FSM Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Ok {
  ok: boolean;
}

export interface ErrorBody {
  error: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface AuthUser {
  id: number;
  email: string;
  fullName: string;
  role: string;
  department: string;
}

export interface AuthMeResponse {
  user: AuthUser;
}

export interface DashboardKpi {
  openJobs: number;
  scheduledJobs: number;
  completedThisMonth: number;
  pipelineCents: number;
  paidThisMonthCents: number;
  outstandingInvoicesCents: number;
  fleetActive: number;
  fleetInShop: number;
  employees: number;
}

export interface DashboardJobByStatus {
  status: string;
  count: number;
}

export interface DashboardRecentJob {
  id: number;
  status: string;
  totalCents: number;
  scheduledFor?: string | null;
  notes?: string | null;
}

export interface DashboardSummary {
  kpi: DashboardKpi;
  jobsByStatus: DashboardJobByStatus[];
  quotePipelineByStage: DashboardJobByStatus[];
  recentJobs: DashboardRecentJob[];
}

export interface Customer {
  id: number;
  fullName: string;
  email?: string | null;
  phone?: string | null;
  billingAddress?: string | null;
  ownerUserId: number;
  createdAt: string;
}

export interface CustomerWriteBody {
  fullName: string;
  email?: string | null;
  phone?: string | null;
  billingAddress?: string | null;
  ownerUserId: number;
}

export interface CustomerResponse {
  customer: Customer;
}

export interface CustomerListResponse {
  customers: Customer[];
}

export interface Job {
  id: number;
  propertyId: number;
  crewId?: number | null;
  status: string;
  scheduledFor?: string | null;
  completedAt?: string | null;
  totalCents: number;
  notes?: string | null;
  createdAt: string;
}

export interface JobWriteBody {
  propertyId: number;
  crewId?: number | null;
  status: string;
  scheduledFor?: string | null;
  completedAt?: string | null;
  totalCents: number;
  notes?: string | null;
}

export interface JobResponse {
  job: Job;
}

export interface JobListResponse {
  jobs: Job[];
}

export interface Quote {
  id: number;
  customerId: number;
  propertyId?: number | null;
  ownerUserId: number;
  status: string;
  subtotalCents: number;
  totalCents: number;
  sentAt?: string | null;
  decidedAt?: string | null;
  createdAt: string;
}

export interface QuoteWriteBody {
  customerId: number;
  propertyId?: number | null;
  ownerUserId: number;
  status: string;
  subtotalCents: number;
  totalCents: number;
  sentAt?: string | null;
  decidedAt?: string | null;
}

export interface QuoteResponse {
  quote: Quote;
}

export interface QuoteListResponse {
  quotes: Quote[];
}

export interface Invoice {
  id: number;
  jobId?: number | null;
  customerId: number;
  status: string;
  totalCents: number;
  issuedAt?: string | null;
  paidAt?: string | null;
}

export interface InvoiceWriteBody {
  jobId?: number | null;
  customerId: number;
  status: string;
  totalCents: number;
  issuedAt?: string | null;
  paidAt?: string | null;
}

export interface InvoiceResponse {
  invoice: Invoice;
}

export interface InvoiceListResponse {
  invoices: Invoice[];
}

export interface Truck {
  id: number;
  name: string;
  vin?: string | null;
  plate?: string | null;
  status: string;
  assignedCrewId?: number | null;
}

export interface TruckWriteBody {
  name: string;
  vin?: string | null;
  plate?: string | null;
  status: string;
  assignedCrewId?: number | null;
}

export interface TruckResponse {
  truck: Truck;
}

export interface TruckListResponse {
  trucks: Truck[];
}

export interface Equipment {
  id: number;
  name: string;
  type: string;
  serial?: string | null;
  status: string;
  assignedTruckId?: number | null;
}

export interface EquipmentWriteBody {
  name: string;
  type: string;
  serial?: string | null;
  status: string;
  assignedTruckId?: number | null;
}

export interface EquipmentResponse {
  equipment: Equipment;
}

export interface EquipmentListResponse {
  equipment: Equipment[];
}

export interface MaintenanceLog {
  id: number;
  truckId?: number | null;
  equipmentId?: number | null;
  kind: string;
  description: string;
  performedByUserId?: number | null;
  performedAt: string;
  costCents: number;
}

export interface MaintenanceLogWriteBody {
  truckId?: number | null;
  equipmentId?: number | null;
  kind: string;
  description: string;
  performedByUserId?: number | null;
  performedAt?: string | null;
  costCents: number;
}

export interface MaintenanceLogResponse {
  log: MaintenanceLog;
}

export interface MaintenanceLogListResponse {
  logs: MaintenanceLog[];
}

export interface Employee {
  id: number;
  email: string;
  fullName: string;
  role: string;
  roleId?: number;
  department: string;
  departmentId?: number;
  isActive: boolean;
  createdAt: string;
}

export interface EmployeeCreateBody {
  email: string;
  fullName: string;
  /** @minLength 6 */
  password: string;
  roleId: number;
  departmentId: number;
  isActive?: boolean;
}

export interface EmployeeUpdateBody {
  email?: string;
  fullName?: string;
  password?: string;
  roleId?: number;
  departmentId?: number;
  isActive?: boolean;
}

export interface EmployeeResponse {
  employee: Employee;
}

export interface EmployeeListResponse {
  employees: Employee[];
}

export interface Role {
  id: number;
  key: string;
  label: string;
}

export interface RoleListResponse {
  roles: Role[];
}

export interface Department {
  id: number;
  key: string;
  label: string;
}

export interface DepartmentListResponse {
  departments: Department[];
}

export interface SectionPermission {
  id?: number | null;
  roleKey: string;
  sectionKey: string;
  canView: boolean;
  canEdit: boolean;
  isOverride?: boolean;
}

export interface SectionPermissionListResponse {
  permissions: SectionPermission[];
}

export interface SectionPermissionUpdateBody {
  roleKey: string;
  sectionKey: string;
  canView: boolean;
  canEdit: boolean;
}

export interface PermissionMatrixCell {
  roleKey: string;
  sectionKey: string;
  canView: boolean;
  canEdit: boolean;
  isOverride: boolean;
}

export interface PermissionMatrixResponse {
  roles: Role[];
  sections: string[];
  cells: PermissionMatrixCell[];
}

/**
 * Error
 */
export type ErrorResponse = ErrorBody;