# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: FSM Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CustomerListResponse" }
    post:
      operationId: createCustomer
      tags: [customers]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CustomerWriteBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CustomerResponse" }
  /customers/{id}:
    patch:
      operationId: updateCustomer
      tags: [customers]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CustomerWriteBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CustomerResponse" }
    delete:
      operationId: deleteCustomer
      tags: [customers]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
...
  /admin/users:
    get:
      operationId: listEmployees
      tags: [employees]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EmployeeListResponse" }
    post:
      operationId: createEmployee
      tags: [employees]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EmployeeCreateBody" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EmployeeResponse" }
  /admin/users/{id}:
    patch:
      operationId: updateEmployee
      tags: [employees]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EmployeeUpdateBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EmployeeResponse" }
    delete:
      operationId: deleteEmployee
      tags: [employees]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
...
  /admin/section-permissions:
    get:
      operationId: listSectionPermissions
      tags: [permissions]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SectionPermissionListResponse" }
    post:
      operationId: updateSectionPermission
      tags: [permissions]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SectionPermissionUpdateBody" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ok" }
...
components:
...
  schemas:
...
    Ok:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
...
    Customer:
      type: object
      required: [id, fullName, ownerUserId, createdAt]
      properties:
        id: { type: integer }
        fullName: { type: string }
        email: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        billingAddress: { type: string, nullable: true }
        ownerUserId: { type: integer }
        createdAt: { type: string, format: date-time }
    CustomerWriteBody:
      type: object
      required: [fullName, ownerUserId]
      properties:
        fullName: { type: string }
        email: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        billingAddress: { type: string, nullable: true }
        ownerUserId: { type: integer }
    CustomerResponse:
      type: object
      required: [customer]
      properties:
        customer: { $ref: "#/components/schemas/Customer" }
    CustomerListResponse:
      type: object
      required: [customers]
      properties:
        customers:
          type: array
          items: { $ref: "#/components/schemas/Customer" }
...
    Employee:
      type: object
      required: [id, email, fullName, role, department, isActive, createdAt]
      properties:
        id: { type: integer }
        email: { type: string }
        fullName: { type: string }
        role: { type: string }
        roleId: { type: integer }
        department: { type: string }
        departmentId: { type: integer }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
    EmployeeCreateBody:
      type: object
      required: [email, fullName, password, roleId, departmentId]
      properties:
        email: { type: string, format: email }
        fullName: { type: string }
        password: { type: string, minLength: 6 }
        roleId: { type: integer }
        departmentId: { type: integer }
        isActive: { type: boolean }
    EmployeeUpdateBody:
      type: object
      properties:
        email: { type: string }
        fullName: { type: string }
        password: { type: string }
        roleId: { type: integer }
        departmentId: { type: integer }
        isActive: { type: boolean }
    EmployeeResponse:
      type: object
      required: [employee]
      properties:
        employee: { $ref: "#/components/schemas/Employee" }
    EmployeeListResponse:
      type: object
      required: [employees]
      properties:
        employees:
          type: array
          items: { $ref: "#/components/schemas/Employee" }
...
    SectionPermission:
      type: object
      required: [roleKey, sectionKey, canView, canEdit]
      properties:
        id: { type: integer, nullable: true }
        roleKey: { type: string }
        sectionKey: { type: string }
        canView: { type: boolean }
        canEdit: { type: boolean }
        isOverride: { type: boolean }
    SectionPermissionListResponse:
      type: object
      required: [permissions]
      properties:
        permissions:
          type: array
          items: { $ref: "#/components/schemas/SectionPermission" }
    SectionPermissionUpdateBody:
      type: object
      required: [roleKey, sectionKey, canView, canEdit]
      properties:
        roleKey: { type: string }
        sectionKey: { type: string }
        canView: { type: boolean }
        canEdit: { type: boolean }
...