# lib/api-zod/src/generated/types/customerWriteBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FSM Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CustomerWriteBody {
  fullName: string;
  email?: string | null;
  phone?: string | null;
  billingAddress?: string | null;
  ownerUserId: number;
}