# lib/api-zod/src/generated/types/employee.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FSM Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Employee {
  id: number;
  email: string;
  fullName: string;
  role: string;
  roleId?: number;
  department: string;
  departmentId?: number;
  isActive: boolean;
  createdAt: Date;
}