# lib/api-zod/src/generated/types/employeeCreateBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FSM Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface EmployeeCreateBody {
  email: string;
  fullName: string;
  /** @minLength 6 */
  password: string;
  roleId: number;
  departmentId: number;
  isActive?: boolean;
}